Push CFBData structs to new library
