+++ cfbdata/http.go
@@ -1,0 +1,29 @@
+package cfbdata
+
+import (
+	"fmt"
+	"io"
+	"net/http"
+)
+
+func doRequest(client *http.Client, key string, url string) ([]byte, error) {
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed to build request: %v", err)
+	}
+	req.Header.Add("accept", "application/json")
+	req.Header.Add("Authorization", "Bearer "+key)
+
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, fmt.Errorf("failed to do request: %v", err)
+	}
+	defer resp.Body.Close()
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read response body: %v", err)
+	}
+
+	return body, nil
+}
